refactor: domain events removed from mongo module

+++ src/GoodReads.Domain/RatingAggregate/ValueObjects/Score.cs
@@ -7,9 +7,9 @@ namespace GoodReads.Domain.RatingAggregate.ValueObjects
 {
     public sealed class Score : ValueObject
     {
-        public decimal Value { get; private set; }
+        public int Value { get; private set; }
 
-        public Score(decimal value)
+        public Score(int value)
         {
             value.Throw(() => new DomainException("'Score' must be between 1 and 5"))
                 .IfLessThan(1)
